[ALPS03411241] binder: protect transaction_stack with inner lock.
(from https://patchwork.kernel.org/patch/9817779/)

This makes future changes to priority inheritance
easier, since we want to be able to look at a threads
transaction stack when selecting a thread to inherit
priority for.

It also allows us to take just a single lock in a
few paths, where we used to take two in succession.

MTK-Commit-Id: 31738446cbe36e97bc94ebe1573e9097ac8b3c9b

Change-Id: Idb1b6e9faa5c669978b2b3011fe326be8aece586
Signed-off-by: Martijn Coenen <maco@google.com>
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>
CR-Id: ALPS03411241
Feature: binder