[ALPS03423149] selinux: quick selinux support for tracefs
[Detail]
backport from commit:3a343a1540d4 to resolve CTS failure
due to labeling failure for debugfs_trace_marker...etc.

quick selinux support for tracefs

Here is just the quick fix for tracefs with selinux.
just add tracefs to the list of whitelisted filesystem
types in selinux_is_sblabel_mnt(), but the right fix would be to
generalize this logic as described in the last item on the todo list,
https://bitbucket.org/seandroid/wiki/wiki/ToDo

MTK-Commit-Id: d0fedb8a858a5ee26f46d0dc9194f520dfb63ce5

Change-Id: I4b2e9b196a7271dac59beb53b4fe85a474658862
CR-Id: ALPS03423149
Feature: [Android Default] SELinux, SEAndroid, and SE-MTK
Signed-off-by: Shangbing Hu <shangbing.hu@mediatek.com>