[ALPS03727915] Audio: google security patch
[Detail]
In the MIDI driver global variables in certain ioctl calls
are not correctly protected with locks.
This could lead to double-free or use-after-free errors in the kernel.
[Solution]
The fix is designed to fix the locking in the affected functions.

MTK-Commit-Id: 201f46fe4e13a2772af62eed056ae30caabf6868

Change-Id: I5afd7e46076e2c75e8f3b9a43e95fa28487dcf78
Signed-off-by: chipeng chang <chipeng.chang@mediatek.com>
CR-Id: ALPS03727915
Feature: Others
(cherry-pick from b199b2a3f4202ab735e846554392f825a6996b36)