[ALPS03727915] Audio: google security patch
[Detail]
In the MIDI driver global variables in certain ioctl calls
are not correctly protected with locks.
This could lead to double-free or use-after-free errors in the kernel.
[Solution]
The fix is designed to fix the locking in the affected functions.

MTK-Commit-Id: 5b7e8d79293f51aaa2ad44174497fe7ed67f41c7

Change-Id: I6b75b929a5cf9d58d74dd24b96702006bfc3a2f8
Signed-off-by: chipeng chang <chipeng.chang@mediatek.com>
CR-Id: ALPS03727915
Feature: Others
(cherry-pick from e1c46e391670c66184929563670a9d8c7e1d3840)