USB: core: only clean up what we allocated
[Detail]
When cleaning up the configurations, make sure we only free the number
of configurations and interfaces that we could have allocated.

In usb_destroy_configuration of config.c, there is a possible out of
bounds write due to a missing bounds check. This could lead to local
escalation of privilege with no additional execution privileges needed.
User interaction is not needed for exploitation.

[Solution]
The fix is designed to make sure the driver only frees as many
configurations and interfaces as it could have allocated.

MTK-Commit-Id: 8cd5127cc5ce70b2badd95bfb7e4c2874337b7c9

Change-Id: Ic659b3c24aae1ff6e13dd2d486f060a02e64aba7
Signed-off-by: Eddie Hung <eddie.hung@mediatek.com>
CR-Id: ALPS03913650
Feature: Google Security Patch
(cherry-pick from eccf1ce0aee981417a943356ae6c76891f3dab59)